requirements.txt
alembic==1.14.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
Faker==33.1.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.8
MarkupSafe==2.1.5
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==1.4.23
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==2.0.1
zipp==3.20.2
